Create bower dependency list with bootstrap and jquery.
This works on issue #6.

I also created a Makefile along with this change. It should be noted instead of
installing things with bower through Makefile targets we should be careful to
add them to bower.json as dependencies. That way we only need to run bower
install to get the static css and js components. This saves time and work.

I've left out chartist because we haven't gotten to graphing and that's a
distraction from core elements right now.