Add field and form base classes.
The field class provides an __init__ and a method called evaluate to use.
It might be interesting to look at metamethods for an alternative name.

It should also be noted that work on the Form base class has not been completed yet. The plan is that the form will encapsulate a bunch of field objects and
somehow provide coordinates for them (hard coded or dynamically created).

Have fun!